Tests/reorganize: move addStringEmbeddedImage tests to own file
